pyproject.toml
[project]
name = "zabbix-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "uvicorn>=0.34.1",
]
